CLIHelper_Backend/requirements.txt
flask
flask-cors
flask-limiter
google-generativeai
gunicorn
python-dotenv
